@@ -1,22 +1,22 @@
 name: message_builder
-version: 0.1.7-dev
+version: 0.2.0
 description: >-
   Generate immutable Dart classes serializable to json from a schema.
 homepage: https://github.com/natebosch/message_builder
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.19.0 <3.0.0"
 
 dependencies:
-  build: ^1.1.0
-  build_config: ^0.4.0
-  code_builder: ^3.1.0
-  dart_style: ^1.1.0
-  yaml: ^2.1.0
+  build: ^2.3.0
+  build_config: ^1.0.0
+  code_builder: ^4.6.0
+  dart_style: ^2.2.3
+  yaml: ^3.1.0
 
 dev_dependencies:
-  build_runner: ^1.0.0
-  build_test: ^1.0.0
+  build_runner: ^2.0.0
+  build_test: ^2.0.0
   path: ^1.5.0
-  pedantic: ^1.9.0
+  lints: ">=1.0.0 <3.0.0" #^2.1.0
   test: ^1.13.0
