@@ -1,3 +1,4 @@
+@TestOn('vm')
 import 'dart:io';
 
 import 'package:test/test.dart';
@@ -13,11 +14,11 @@ void main() {
     final goldenDir = Directory('test/goldens/');
     for (final item in goldenDir.listSync()) {
       if (!item.path.endsWith('.yaml') || item is! File) continue;
-      final file = item as File;
+      final file = item;
       test(p.basename(file.path), () {
         final yamlContent = file.readAsStringSync();
         final dartFile =
-            File(file.path.substring(0, file.path.length - 4) + 'dart');
+            File('${file.path.substring(0, file.path.length - 4)}dart');
         final dartContent = dartFile.readAsStringSync();
         return testBuilder(builder, {
           'a|${p.basename(file.path)}': yamlContent
